Time: 28 ms (27.43%), Space: 5.9 MB (72.47%) - LeetHub

+++ 326-power-of-three/326-power-of-three.cpp
@@ -3,7 +3,7 @@ public:
     bool isPowerOfThree(int n) {
         if(n<=0)
             return false;
-        long int l = log10(n)/log10(3);
+        int l = log10(n)/log10(3);
         if(pow(3,l)==n)
             return true;
         return false;
